refactor(workspace): extract the client projection into its own module
Adding getWorkspaceForClient in 42e63e81 pushed
app/lib/workspace-members-db.server.ts from 786 to 823 lines, over the 800-line
ratchet in tools:check-file-size. That gate is not on ci:local's happy path
locally, so it was caught late -- the regression was mine.

Move the projection to app/lib/workspace-client-projection.server.ts rather
than allowlist the file: the ratchet only ratchets down, and a client-safe
column projection is a distinct concern from member CRUD anyway. members-db is
back to its original 786 lines.

getWorkspaceById stays in workspace-members-db.server.ts for server-only
callers, so check:workspace-projection is unaffected.
