+++ app/routes/workspaces+/$id/audios/record.loader.server.ts
@@ -1,7 +1,7 @@
 import { data as routeData } from "react-router";
 
 import { defineLoader } from "@/lib/handler.server";
-import { getWorkspaceForClient } from "@/lib/workspace-members-db.server";
+import { getWorkspaceForClient } from "@/lib/workspace-client-projection.server";
 import { workspaceLoaderAuth } from "@/lib/workspace-route.server";
 
 export const loader = defineLoader({
